pyproject.toml
[project]
name = "vehicle-analyzer"
version = "0.1.0"
description = "API for vehicle detection and image description"
authors = [
    {name = "Your Name"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "ultralytics (>=8.3.127,<9.0.0)",
    "torchvision (>=0.22.0,<0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.3.5"

